Refactor: Change directory name from components to projects
