+++ src/pages/TodoListPage.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import TodoList from '../components/todo-list/TodoList';
+import TodoList from '../projects/todo-list/TodoList';
 
 export default function TodoListPage() {
   return (
