+++ src/projects/rating/Rating.tsx
