+++ src/projects/rating/RatingInput.tsx
