+++ src/projects/rating/RatingStyle.ts
