+++ src/projects/rating/constants.ts
