+++ src/projects/rating/type.ts
