+++ src/projects/rating/util.ts
